增加IM模块-用户表/用户与消息关系表,修改表ID

+++ src/main/java/vip/qsos/ktorm/module/chat/entity/ChatGroup.kt
@@ -12,7 +12,7 @@ package vip.qsos.ktorm.module.chat.entity
  */
 data class ChatGroup(
         /** @see ChatSession.sessionId */
-        val groupId: Long,
+        val groupId: Int,
         val name: String,
         val createTime: Long,
         val notice: String? = null
